Implemented class Data as a base class for Events, Made Event inherit from Data, changed Tests accordingly

+++ src/DataSet.cpp
@@ -71,7 +71,7 @@ DataSet::DataSet(const DataSet& original)
 
 	//deep copy
 	//note: range based for (aka for each) does not work, since that would be a copy of the unique pointer
-	for(int i = 0; i < original.getSize(); i++)
+	for(size_t i = 0; i < original.getSize(); i++)
 	{
 		unique_ptr<Event> temp(new Event(original[i]));
